% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsample-package.R, R/adsample.R
\docType{package}
\name{adsample}
\alias{adsample}
\alias{adsample-package}
\alias{adsample}
\alias{plot.adsample}
\title{adsample: Adaptive sampling routine for log concave densities}
\usage{
adsample(n, log_dens, initialPoints, minRange, maxRange, debug = FALSE, ...)

\method{plot}{adsample}(x, ...)
}
\arguments{
\item{n}{number of variates to draw}

\item{log_dens}{log density and its derivative, a function with a single
parameter}

\item{initialPoints}{vector of initial points within the domain of log_dens}

\item{minRange}{lower boundary of support of log_dens}

\item{maxRange}{upper boundary of support of log_dens}

\item{debug}{if true, return a list of the algorithm's internal state}

\item{...}{additional parameters to pass to the log density function}

\item{x}{an adsample object}
}
\value{
a vector of samples, or if debug is TRUE, a list of samples and
        state variables
}
\description{
adsample: Adaptive sampling routine for log concave densities

If debug is TRUE, returns a list rather than a vector of samples

Create an adsample object by calling adsample with debug=TRUE.
This function plots empirical and actual density side by side.
}
\references{
Gilks, W. R., & Wild, P. (1992). Adaptive rejection sampling
            for Gibbs sampling. Applied Statistics, 337–348.
}
\seealso{
mklogdensf
}
