% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsample.R
\name{mklogdensf}
\alias{mklogdensf}
\title{Compute a symbolic log density function for use with \code{adsample}.}
\usage{
mklogdensf(expr, wrt, ...)
}
\arguments{
\item{expr}{expression proportional to the density (not log density)}

\item{wrt}{derivative with respect to this variable}

\item{...}{additional parameters to fix in density function}
}
\value{
a function suitable for use with adsample
}
\description{
You must specify a *single* variable to differentiate with respect to.
\code{mklogensf} returns a function from the abscissa into a bivariate vector
of the form \code{( log(density(x)), dlog(density(x))/dx )}.
}
\examples{

# generated function takes any additional parameters
g <- mklogdensf(a + b^2, b)
g(3, a = 2)

# you can also fix additional parameters when you call mklogdensf
g <- mklogdensf(a + b^2, b, a = 2)
g(3)
g(4)

}
