% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raw_ad_sample}
\alias{raw_ad_sample}
\title{Rcpp wrapper for adsamp algorithm}
\usage{
raw_ad_sample(n, log_dens, initialPoints, minRange, maxRange)
}
\arguments{
\item{n}{number of variates to draw}

\item{log_dens}{log density function, a function with a single parameter}

\item{initialPoints}{at least 2 points in support to seed algorithm}

\item{minRange}{lower bound of support}

\item{maxRange}{upper bound of support}
}
\description{
Rcpp wrapper for adsamp algorithm
}
