% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raw_ad_sample_debug}
\alias{raw_ad_sample_debug}
\title{Duplicate wrapper for adsamp algorithm for debugging. This version
returns the algorithm state as well as the variates.}
\usage{
raw_ad_sample_debug(n, log_dens, initialPoints, minRange, maxRange, maxiter)
}
\arguments{
\item{n}{number of variates to draw}

\item{log_dens}{log density function, a function with a single parameter}

\item{initialPoints}{at least 2 points in support to seed algorithm}

\item{minRange}{lower bound of support}

\item{maxRange}{upper bound of support}

\item{maxiter}{maximum number of iterations; zero if no limit}
}
\description{
Duplicate wrapper for adsamp algorithm for debugging. This version
returns the algorithm state as well as the variates.
}
