% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsample-package.R, R/adsample.R
\docType{package}
\name{adsample}
\alias{adsample}
\alias{adsample-package}
\alias{adsample}
\alias{plot.adsample}
\title{adsample: Adaptive rejection sampling for log concave densities}
\usage{
adsample(n, log_dens, initialPoints, minRange, maxRange, debug = FALSE,
  maxiter = 0, ...)

\method{plot}{adsample}(x, ...)
}
\arguments{
\item{n}{number of variates to draw}

\item{log_dens}{log density and its derivative, a function with a single
parameter}

\item{initialPoints}{vector of initial points within the domain of log_dens}

\item{minRange}{lower boundary of support of log_dens}

\item{maxRange}{upper boundary of support of log_dens}

\item{debug}{if true, return a list of the algorithm's internal state}

\item{maxiter}{maximum number of iterations, or zero for no limit}

\item{...}{additional parameters to pass to the log density function}

\item{x}{an adsample object}
}
\value{
a vector of samples, or if debug is TRUE, a list of samples and
        state variables
}
\description{
adsample: Adaptive rejection sampling for log concave densities

If debug is TRUE, returns a list rather than a vector of samples

Create an adsample object by calling adsample with debug=TRUE.
This function plots empirical and actual density side by side.
}
\examples{
\dontrun{
 h <- function(tau) {
    y <- log(4*sqrt(2)*sqrt(tau)*exp(-1/2*tau*(3 - 1)^2)/sqrt(pi))
    yprime <- 1/2*(-tau*(3 - 1)^2 + 1)/tau
    c(y, yprime)
  }
 xs <- adsample(n = 10, log_dens = h, initialPoints = c(0.5, 1),
                minRange = 0, maxRange = Inf)
}
}
\references{
Gilks, W. R., & Wild, P. (1992). Adaptive rejection sampling
            for Gibbs sampling. Applied Statistics, 337–348.
}
\seealso{
mklogdensf
}
